Address remaining PR #779 review feedback
- Apply chain arc damage to AutoTurretDrone targets: add HP system
  (hp/maxHp/takeDamage/resetHp) so drones take damage and deactivate
  at 0 HP. Game loop now calls takeDamage on arc hit targets.
- Change charge_beam fireRateMultiplier from 0.0 to 1.0 to prevent
  latent division-by-zero if routed through generic fire-rate formula.
- Add explicit 'implements BeamLike' to EnemyChargeBeam for type safety.
  Move BeamLike interface from CollisionSystem to shared types.ts to
  avoid entity-to-system import dependency.

Made-with: Cursor
